requirements.txt
# Python dependencies for data import scripts
google-cloud-firestore>=2.11.0
requests>=2.31.0
python-geojson>=3.0.0
pandas>=2.0.0
tqdm>=4.65.0
